Adjust the directory structure
