+++ GatewayService/GatewayServiceServer.cpp
@@ -1,4 +1,4 @@
-#include "GatewayServiceServer.h"
+#include "GatewayService/GatewayServiceServer.h"
 #include <bcos-tars-protocol/Common.h>
 using namespace bcostars;
 bcostars::Error GatewayServiceServer::asyncNotifyGroupInfo(
