+++ NativeNode/main/NativeNodeApp.h
@@ -13,28 +13,33 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  *
- * @brief application for PBFTService
- * @file PBFTServiceApp.h
+ * @brief Application for the NativeNode
+ * @file NativeNodeApp.h
  * @author: yujiechen
- * @date 2021-10-15
+ * @date 2021-10-18
  */
 #pragma once
-#include "../Common/TarsUtils.h"
-#include "libinitializer/PBFTInitializer.h"
-#include "PBFTServiceServer.h"
+#include "Common/TarsUtils.h"
+#include "FrontService/FrontServiceServer.h"
+#include "LedgerService/LedgerServiceServer.h"
+#include "PBFTService/PBFTServiceServer.h"
+#include "SchedulerService/SchedulerServiceServer.h"
+#include "TxPoolService/TxPoolServiceServer.h"
+#include "libinitializer/Initializer.h"
+#include <bcos-framework/libtool/NodeConfig.h>
 #include <bcos-framework/libutilities/BoostLogInitializer.h>
 #include <tarscpp/servant/Application.h>
 
 namespace bcostars
 {
-class PBFTServiceApp : public Application
+class NativeNodeApp : public Application
 {
 public:
-    PBFTServiceApp() {}
-    virtual ~PBFTServiceApp() override{};
+    NativeNodeApp() {}
+    ~NativeNodeApp() override {}
 
     void initialize() override;
-    void destroyApp() override {}
+    void destroyApp() override { m_nodeInitializer->stop(); }
 
 protected:
     virtual void initConfig()
@@ -46,13 +51,19 @@ protected:
         addConfig("config.genesis");
         addConfig("config.ini");
     }
-    virtual void initService();
+    virtual void initLog();
+    virtual void initNodeService();
+    virtual void initTarsNodeService();
+    void initHandler();
+    void notifyBlockNumberToAllRpcNodes(bcostars::RpcServicePrx _rpcPrx,
+        bcos::protocol::BlockNumber _blockNumber, std::function<void(bcos::Error::Ptr)> _callback);
 
 private:
     std::string m_iniConfigPath;
     std::string m_genesisConfigPath;
     std::string m_privateKeyPath;
+
     bcos::BoostLogInitializer::Ptr m_logInitializer;
-    bcos::initializer::PBFTInitializer::Ptr m_pbftInitializer;
+    bcos::initializer::Initializer::Ptr m_nodeInitializer;
 };
 }  // namespace bcostars
