+++ RpcService/main/Application.cpp
@@ -1,6 +1,6 @@
-#include "../Common/TarsUtils.h"
-#include "RpcInitializer.h"
-#include "RpcServiceServer.h"
+#include "Common/TarsUtils.h"
+#include "RpcService/RpcInitializer.h"
+#include "RpcService/RpcServiceServer.h"
 #include <bcos-framework/libutilities/BoostLogInitializer.h>
 #include <tarscpp/servant/Application.h>
