+++ TxPoolService/TxPoolServiceServer.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "../Common/TarsUtils.h"
-#include "../libinitializer/ProtocolInitializer.h"
+#include "Common/TarsUtils.h"
+#include "libinitializer/ProtocolInitializer.h"
 #include "libinitializer/TxPoolInitializer.h"
 #include <bcos-framework/interfaces/consensus/ConsensusNode.h>
 #include <bcos-framework/libutilities/Common.h>
