+++ libinitializer/FrontServiceInitializer.h
@@ -19,8 +19,8 @@
  * @date 2021-06-10
  */
 #pragma once
-#include "../Common/TarsUtils.h"
-#include "../libinitializer/ProtocolInitializer.h"
+#include "Common/TarsUtils.h"
+#include "libinitializer/ProtocolInitializer.h"
 #include <bcos-framework/interfaces/consensus/ConsensusInterface.h>
 #include <bcos-framework/interfaces/gateway/GatewayInterface.h>
 #include <bcos-framework/interfaces/sync/BlockSyncInterface.h>
