+++ libinitializer/LedgerInitializer.h
@@ -36,8 +36,8 @@ public:
     {
         auto ledger = std::make_shared<bcos::ledger::Ledger>(_blockFactory, _storage);
         // build genesis block
-        ledger->buildGenesisBlock(_nodeConfig->ledgerConfig(), _nodeConfig->groupId(),
-            _nodeConfig->txGasLimit(), _nodeConfig->genesisData());
+        ledger->buildGenesisBlock(
+            _nodeConfig->ledgerConfig(), _nodeConfig->txGasLimit(), _nodeConfig->genesisData());
         return ledger;
     }
 };
